// src/shared/api/model/perfilResponse.ts
/**
 * Generated by orval v8.17.0 🍺
 * Do not edit manually.
 * API de Gestão de Portfólio
 * Contrato de API para o CMS e exibição da Landing Page do Portfólio (Modulo Portfolio)
 * OpenAPI spec version: 1.0.0
 */

export interface PerfilResponse {
  id?: string;
  nome?: string;
  titulo?: string;
  bio?: string;
  linkLinkedin?: string;
  linkGithub?: string;
  urlFoto?: string;
  usuarioId?: string;
  versao?: number;
}